+++ config/config.exs
@@ -19,8 +19,6 @@ config :teiserver,
   # Overridden by application
   client_destroy_timeout_seconds: 300,
   lobby_join_method: :simple,
-  teiserver_clustering: true,
-  teiserver_clustering_post_join_functions: [],
 
   # User defaults
   default_behaviour_score: 10_000,